repository read files should be renamed into getOne, renamed create/removeCycleSchema into create/removeSchema
