+++ server/repository/assessmentCycle/createSchema.ts
@@ -1,17 +1,20 @@
 import { Assessment, Cycle } from '@meta/assessment'
 import { BaseProtocol, DB, Schemas } from '@server/db'
+import { getCreateSchemaCycleDDL } from '@server/repository/assessment/getCreateSchemaDDL'
 
-export const removeCycleSchema = async (
+export const createSchema = async (
   params: {
-    assessment: Assessment
+    assessment: Pick<Assessment, 'props'>
     cycle: Cycle
   },
   client: BaseProtocol = DB
 ): Promise<string> => {
   const { assessment, cycle } = params
+
+  const schemaName = Schemas.getName(assessment)
   const cycleSchemaName = Schemas.getNameCycle(assessment, cycle)
 
-  await client.query<void>(`drop schema ${cycleSchemaName} cascade;`)
+  await client.query(getCreateSchemaCycleDDL(schemaName, cycleSchemaName))
 
   return cycleSchemaName
 }
