+++ server/repository/assessmentCycle/removeSchema.ts
@@ -1,20 +1,17 @@
 import { Assessment, Cycle } from '@meta/assessment'
 import { BaseProtocol, DB, Schemas } from '@server/db'
-import { getCreateSchemaCycleDDL } from '@server/repository/assessment/getCreateSchemaDDL'
 
-export const createCycleSchema = async (
+export const removeSchema = async (
   params: {
-    assessment: Pick<Assessment, 'props'>
+    assessment: Assessment
     cycle: Cycle
   },
   client: BaseProtocol = DB
 ): Promise<string> => {
   const { assessment, cycle } = params
-
-  const schemaName = Schemas.getName(assessment)
   const cycleSchemaName = Schemas.getNameCycle(assessment, cycle)
 
-  await client.query(getCreateSchemaCycleDDL(schemaName, cycleSchemaName))
+  await client.query<void>(`drop schema ${cycleSchemaName} cascade;`)
 
   return cycleSchemaName
 }
